+++ odoo/local-src/sf_sale_invoice_fiscal_position/__manifest__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Part of senseFly.
+
+{
+    'name': 'Sale invoice fiscal position',
+    'version': '1.0',
+    'author': 'Telmo Santos',
+    'website': 'https://www.sensefly.com',
+    'category': 'SenseFly',
+    'depends': ['sale', 'account_invoice_fiscal_position_update'],
+    'data': [],
+    'demo': [],
+    'images': ['static/description/icon.png'],
+    'installable': True,
+    'auto_install': False,
+}